boot: do not import go-efilib on nosecboot
go-efilib always queries for efivarfs on initialization, which is not
needed for most use of snapd. We can remove import of go-efilib
when nosecboot tag is used.
