+++ boot/export_test.go
@@ -23,9 +23,6 @@ import (
 	"fmt"
 	"sync/atomic"
 
-	"github.com/canonical/go-efilib"
-	"github.com/canonical/go-efilib/linux"
-
 	"github.com/snapcore/snapd/asserts"
 	"github.com/snapcore/snapd/bootloader"
 	"github.com/snapcore/snapd/kernel/fde"
@@ -289,34 +286,4 @@ func MockWriteModelToUbuntuBoot(mock func(*asserts.Model) error) (restore func()
 func EnableTestingRebootFunction() (restore func()) {
 	testingRebootItself = true
 	return func() { testingRebootItself = false }
-}
-
-var (
-	ConstructLoadOption      = constructLoadOption
-	SetEfiBootOptionVariable = setEfiBootOptionVariable
-	SetEfiBootOrderVariable  = setEfiBootOrderVariable
-)
-
-func MockEfiListVariables(f func() ([]efi.VariableDescriptor, error)) (restore func()) {
-	restore = testutil.Backup(&efiListVariables)
-	efiListVariables = f
-	return restore
-}
-
-func MockEfiReadVariable(f func(name string, guid efi.GUID) ([]byte, efi.VariableAttributes, error)) (restore func()) {
-	restore = testutil.Backup(&efiReadVariable)
-	efiReadVariable = f
-	return restore
-}
-
-func MockEfiWriteVariable(f func(name string, guid efi.GUID, attrs efi.VariableAttributes, data []byte) error) (restore func()) {
-	restore = testutil.Backup(&efiWriteVariable)
-	efiWriteVariable = f
-	return restore
-}
-
-func MockLinuxFilePathToDevicePath(f func(path string, mode linux.FilePathToDevicePathMode) (out efi.DevicePath, err error)) (restore func()) {
-	restore = testutil.Backup(&linuxFilePathToDevicePath)
-	linuxFilePathToDevicePath = f
-	return restore
 }
